+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Changes should still be described appropriately in JIRA/doc input pages, for inc
 ### Technical Changes
 - ROX-30769: Update Node.js requirement for ui folder to 22.13.0
 - ROX-31295: The lower limit for `ROX_MAX_PARALLEL_IMAGE_SCAN_INTERNAL` on Sensor has been reduced to one (from 10).
+- ROX-32051: Fixed an issue in the compliance v2 API that could lead to inconsistent data being returned for orphaned profiles of the compliance operator.
 
 ## [4.9.0]